Delete Jekyll theme
This is probably not the best way to do this, but I didn't find a better solution. The Jekyll theme's _site caused a "stashing" problem and couldn't merge my master and origin. I made manual copies of those files that needed update, so I'm deleting and uploading manually the changes (instead of using GitKraken, where I found the problem).